% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flc.R
\name{flc_search}
\alias{flc_search}
\title{Search federal lab inventions by keyword}
\usage{
flc_search(query)
}
\arguments{
\item{query}{Character string to search for (case-insensitive)}
}
\value{
tibble: same columns as flc_inventions()
}
\description{
Search federal lab inventions by keyword
}
